Simplify Containers
Reorganize containers to where values are stored in accordance with how the containers look in Python. Links are then stored according to where the value occurs in the TOML syntax.

Further instead of trying to handle links moving around when a container's complexity changes we insert the link into two locations when it is first created, the root and the local scope. This allows us to only worry about where a link is and whether to display it or to ignore it when we are actually dumping into TOML syntax.